+++ src/model.h
@@ -154,7 +154,7 @@ public:
 
   double likelihood(std::vector<double> &temp_suff_stat, std::vector<double> &suff_stat_all, bool left_side, bool no_split, std::unique_ptr<State> &state) const;
 
-  void predict_std(const double *Xtestpointer, const double *tpointer, size_t N_test, size_t p, size_t num_sweeps, std::vector<matrix<double>> &yhats_test_xinfo, vector<vector<tree>> &trees);
+  void predict_std(const double *Xtestpointer, const double *tpointer, size_t N_test, size_t p, size_t num_sweeps, std::vector<matrix<double>> &yhats_test_xinfo, vector<vector<tree>> &trees, const std::vector<const double *> *tv = 0);
 
   void update_a_value(std::unique_ptr<State> &state);